memory: remove support for address space extension.
The extend= parameter of MemoryMap.add_resource()/.add_window() was
added in 2d18ac7d9fc810c076c74c3bebb6a099e777038b.

However, it prevents interfaces whose shape depend on a memory map from
being instantiated until the latter is frozen. Also, bus interfaces may
have a strict requirement on their address width, and it should not be
possible to implicitly increase it by adding a subordinate resource or
window.

As a side-effect of this change, csr.EventMonitor requires a populated
EventMap in order to be instantiated.