csr.reg: move access= parameter to Register.__init_subclass__().
This allows csr.Register subclasses that use variable annotations to
define their access mode without overriding __init__().

In addition, the default value of access= (which was "rw") is removed.
The access mode of a register has an influence on the rest of the SoC
and should therefore be explicitly set.

As a consequence of access= being a required __init_subclass__() kwarg,
the csr.Register class can no longer be directly instantiated and must
be subclassed first.